Move subcomponents of BaseLayerRightToolbar into their own files.
