+++ frontend/src/features/map_planning/layers/base/components/BaseLayerRightToolbar.tsx
@@ -1,51 +1,19 @@
 import { UpdateBaseLayerAction } from '../../../layers/base/actions';
 import { BaseLayerImageDto } from '@/api_types/definitions';
-import SimpleButton from '@/components/Button/SimpleButton';
-import SimpleFormInput from '@/components/Form/SimpleFormInput';
-import ModalContainer from '@/components/Modals/ModalContainer';
+import {
+  BaseLayerAttributeEditForm,
+  BaseLayerAttributeEditFormDataAttributes,
+} from '@/features/map_planning/layers/base/components/BaseLayerAttributeEditForm';
+import {
+  BaseLayerDistanceModalAttributes,
+  DistanceMeasurementModal,
+} from '@/features/map_planning/layers/base/components/DistanceMeasurementModal';
 import { calculateDistance, calculateScale } from '@/features/map_planning/layers/base/util';
 import useMapStore from '@/features/map_planning/store/MapStore';
 import { useIsReadOnlyMode } from '@/features/map_planning/utils/ReadOnlyModeContext';
-import FileSelectorModal from '@/features/nextcloud_integration/components/FileSelectorModal';
-import { useDebouncedSubmit } from '@/hooks/useDebouncedSubmit';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { useState } from 'react';
-import { SubmitHandler, useForm } from 'react-hook-form';
+import { SubmitHandler } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { toast } from 'react-toastify';
-import { FileStat } from 'webdav';
-import { z } from 'zod';
-
-export type BaseLayerDataAttributes = Pick<BaseLayerImageDto, 'path' | 'rotation' | 'scale'>;
-
-const BaseLayerRightToolbarFormSchema = z.object({
-  path: z.nullable(z.string()).transform((value) => value ?? ''),
-  rotation: z.number().transform((value) => value % 360),
-  scale: z.number().min(0),
-});
-
-export type BaseLayerDistanceModalAttributes = {
-  meters: number;
-  centimeters: number;
-};
-
-const BaseLayerDistanceModalFormSchema = z.object({
-  meters: z
-    .number()
-    .min(0)
-    .transform((value) => value),
-  centimeters: z
-    .number()
-    .min(0)
-    .max(99)
-    .transform((value) => value),
-});
-
-export const TEST_IDS = Object.freeze({
-  BACKGROUND_INPUT: 'base-layer-right-toolbar__background-input',
-  ROTATION_INPUT: 'base-layer-right-toolbar__rotation-input',
-  SCALE_INPUT: 'base-layer-right-toolbar__scale-input',
-});
 
 class ValidationError extends Error {
   constructor(msg: string) {
@@ -94,7 +62,11 @@ export const BaseLayerRightToolbar = () => {
   const deactivateMeasurement = useMapStore((state) => state.baseLayerDeactivateMeasurement);
   const clearStatusPanelContent = useMapStore((state) => state.clearStatusPanelContent);
 
-  const onBaseLayerFormChange = ({ scale, rotation, path }: BaseLayerDataAttributes) => {
+  const onBaseLayerFormChange = ({
+    scale,
+    rotation,
+    path,
+  }: BaseLayerAttributeEditFormDataAttributes) => {
     const baseLayerOptions = {
       id: baseLayerState.imageId,
       layer_id: baseLayerState.layerId,
@@ -143,7 +115,7 @@ export const BaseLayerRightToolbar = () => {
         onCancel={deactivateMeasurement}
         show={measureStep === 'both selected'}
       />
-      <BaseLayerEditForm
+      <BaseLayerAttributeEditForm
         // remount the form when the selected planting or the step changes (on undo/redo)
         key={`${baseLayerState.id}-${step}`}
         onChange={onBaseLayerFormChange}
@@ -152,175 +124,5 @@ export const BaseLayerRightToolbar = () => {
     </div>
   );
 };
-
-interface BaseLayerEditFormProps {
-  onChange: (data: BaseLayerDataAttributes) => void;
-  isReadOnlyMode: boolean;
-}
-
-function BaseLayerEditForm({ onChange, isReadOnlyMode }: BaseLayerEditFormProps) {
-  const { t } = useTranslation(['common', 'baseLayerForm']);
-
-  const activateMeasurement = useMapStore((state) => state.baseLayerActivateMeasurement);
-  const deactivateMeasurement = useMapStore((state) => state.baseLayerDeactivateMeasurement);
-  const setStatusPanelContent = useMapStore((state) => state.setStatusPanelContent);
-  const clearStatusPanelContent = useMapStore((state) => state.clearStatusPanelContent);
-  const measureStep = useMapStore((state) => state.untrackedState.layers.base.measureStep);
-  const baseLayerState = useMapStore((state) => state.trackedState.layers.base);
-
-  const { register, handleSubmit, watch, setValue, formState } = useForm<BaseLayerDataAttributes>({
-    values: {
-      scale: baseLayerState.scale,
-      rotation: baseLayerState.rotation,
-      path: baseLayerState.nextcloudImagePath,
-    },
-    resolver: zodResolver(BaseLayerRightToolbarFormSchema),
-  });
-
-  useDebouncedSubmit<BaseLayerDataAttributes>(watch('scale'), handleSubmit, onChange);
-  useDebouncedSubmit<BaseLayerDataAttributes>(watch('rotation'), handleSubmit, onChange);
-  useDebouncedSubmit<BaseLayerDataAttributes>(watch('path'), handleSubmit, onChange);
-
-  const [showFileSelector, setShowFileSelector] = useState(false);
-
-  return (
-    <div className="flex flex-col gap-4">
-      <h2>{t('baseLayerForm:title')}</h2>
-      <FileSelectorModal
-        setShow={function (show: boolean): void {
-          setShowFileSelector(show);
-        }}
-        show={showFileSelector}
-        onCancel={function (): void {
-          setShowFileSelector(false);
-        }}
-        path={'/Photos/'}
-        onSelect={function (item: FileStat): void {
-          const path = '/Photos/' + item.basename;
-          setValue('path', path);
-          setShowFileSelector(false);
-        }}
-        title={t('baseLayerForm:selectImage')}
-      />
-      <div className="flex flex-col gap-1">
-        <SimpleFormInput
-          id="path"
-          disabled={isReadOnlyMode}
-          labelText={t('baseLayerForm:image_path_field')}
-          data-testid={TEST_IDS.BACKGROUND_INPUT}
-          register={register}
-        />
-
-        <SimpleButton onClick={() => setShowFileSelector(true)} disabled={isReadOnlyMode}>
-          {t('baseLayerForm:selectImage')}
-        </SimpleButton>
-      </div>
-      <SimpleFormInput
-        id="rotation"
-        register={register}
-        disabled={isReadOnlyMode}
-        labelText={t('baseLayerForm:rotation_field')}
-        type="number"
-        data-testid={TEST_IDS.ROTATION_INPUT}
-      />
-      <div className="flex flex-col gap-1">
-        <SimpleFormInput
-          id="scale"
-          register={register}
-          disabled={isReadOnlyMode}
-          labelText={t('baseLayerForm:scale')}
-          type="number"
-          data-testid={TEST_IDS.SCALE_INPUT}
-        />
-        {formState.errors.scale && (
-          <div className="text-sm text-red-400">{t('baseLayerForm:scale_invalid')}</div>
-        )}
-        {measureStep === 'inactive' ? (
-          <SimpleButton
-            disabled={isReadOnlyMode}
-            onClick={() => {
-              activateMeasurement();
-              setStatusPanelContent(<span>{t('baseLayerForm:auto_scaling_hint')}</span>);
-            }}
-          >
-            {t('baseLayerForm:set_scale')}
-          </SimpleButton>
-        ) : (
-          <SimpleButton
-            onClick={() => {
-              deactivateMeasurement();
-              clearStatusPanelContent();
-            }}
-          >
-            {t('common:cancel')}
-          </SimpleButton>
-        )}
-      </div>
-    </div>
-  );
-}
-
-interface DistanceMeasurementModalProps {
-  onSubmit: SubmitHandler<BaseLayerDistanceModalAttributes>;
-  onCancel: () => void;
-  show: boolean;
-}
-
-function DistanceMeasurementModal({ onSubmit, onCancel, show }: DistanceMeasurementModalProps) {
-  const { t } = useTranslation(['common', 'baseLayerForm']);
-
-  const {
-    register: registerDistanceModal,
-    handleSubmit: handleSubmitDistanceModal,
-    formState: registerDistanceModalFormState,
-  } = useForm<BaseLayerDistanceModalAttributes>({
-    defaultValues: {
-      meters: 0,
-      centimeters: 0,
-    },
-    resolver: zodResolver(BaseLayerDistanceModalFormSchema),
-  });
-
-  return (
-    <ModalContainer show={show}>
-      <div className="w-ful flex h-full min-h-[20vh] flex-col gap-2 rounded-lg bg-neutral-100 p-6 dark:bg-neutral-100-dark">
-        <h3>{t('baseLayerForm:distance_modal_title')}</h3>
-        {registerDistanceModalFormState.errors.meters && (
-          <div className="text-sm text-red-400">
-            {t('baseLayerForm:auto_scaling_meters_invalid')}
-          </div>
-        )}
-        {registerDistanceModalFormState.errors.centimeters && (
-          <div className="text-sm text-red-400">
-            {t('baseLayerForm:auto_scaling_centimeters_invalid')}
-          </div>
-        )}
-        <div className="grid grid-cols-2 gap-2">
-          <SimpleFormInput
-            id="meters"
-            labelText={t('common:meters')}
-            type="number"
-            register={registerDistanceModal}
-            min="0"
-          />
-          <SimpleFormInput
-            id="centimeters"
-            labelText={t('common:centimeters')}
-            type="number"
-            register={registerDistanceModal}
-            min="0"
-            max="99"
-          />
-        </div>
-        <div className="grid grid-cols-2 gap-2">
-          <SimpleButton onClick={() => onCancel()}>{t('common:cancel')}</SimpleButton>
-          <SimpleButton onClick={handleSubmitDistanceModal(onSubmit)}>
-            {t('common:ok')}
-          </SimpleButton>
-        </div>
-      </div>
-    </ModalContainer>
-  );
-}
 
 export default BaseLayerRightToolbar;
